requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
faiss-cpu==1.13.2
fastapi==0.115.0
filelock==3.25.2
fsspec==2026.3.0
grandalf==0.8
gTTS==2.5.3
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langchain-ollama==0.2.0
langgraph==0.2.45
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.6.8
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
ollama==0.6.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pillow==10.4.0
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.20.0
pyparsing==3.3.2
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2026.3.32
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
starlette==0.38.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
typer==0.9.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.32.0
wheel==0.46.3
xxhash==3.6.0
zstandard==0.25.0
streamlit>=1.35.0
